Groundwork for Ring Scan
Coded methods to begin implementing the ring scanning feature.  This
allows one to specify the distance from the player bot in x and y
distance from the player bot.  We then draw a ring to the cardinal
directions.  From there, extend lines outward perpendicular to these
cardinal directions.  This should form a square box.  From here we have
a square ring in which we will scan those boxes for fruit.  Groundwork
for supporting methods also included in this update.